+++ android/app/src/main/assets/extensions/vscodroid.vscodroid-welcome-1.9.0/templates/dart/bin/main.dart
@@ -1,0 +1,3 @@
+void main() {
+  print('{{PROJECT_NAME}}');
+}