+++ android/app/src/main/assets/extensions/vscodroid.vscodroid-welcome-1.9.0/templates/kotlin/src/main/kotlin/Main.kt
@@ -1,0 +1,3 @@
+fun main() {
+    println("{{PROJECT_NAME}}")
+}