+++ android/app/src/main/assets/extensions/vscodroid.vscodroid-welcome-1.9.0/templates/php/index.php
@@ -1,0 +1,3 @@
+<?php
+
+echo "{{PROJECT_NAME}}\n";