+++ android/app/src/main/assets/extensions/vscodroid.vscodroid-welcome-1.9.0/templates/ruby/app.rb
@@ -1,0 +1,1 @@
+puts "{{PROJECT_NAME}}"