+++ android/app/src/main/assets/extensions/vscodroid.vscodroid-welcome-1.9.0/templates/typescript/src/index.ts
@@ -1,0 +1,3 @@
+const projectName: string = '{{PROJECT_NAME}}';
+
+console.log(projectName);